Fixed CVODES uninitialised memory bug and wrote PBS qsub scripts for Imperial's cluster.

+++ examples/FitzHugh_Benchmark_MH.c
@@ -297,7 +297,7 @@ int main(int argc, char * argv[]) {
 
   // Seed the RNG
   time_t seed = time(NULL);
-  gmcmc_prng64_seed(rng, seed);
+  gmcmc_prng64_seed(rng, (uint64_t)seed);
   fprintf(stdout, "Using PRNG seed: %ld\n", seed);
 
   // Start timer
